[AIRFLOW-5782] Move /providers/aws to /providers/amazon/aws [AIP-21]
[AIRFLOW-5782] Fix broken imports.

[AIRFLOW-5782] Refresh updating.md

[AIRFLOW-5782] Restore temporary pending-deprecation imports.

[AIRFLOW-5782] Update sphinx excluded files.

[AIRFLOW-5782] Fix broken mock patches.
