+++ airflow/contrib/operators/aws_sqs_publish_operator.py
@@ -17,15 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""This module is deprecated. Please use `airflow.providers.aws.operators.sqs`."""
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.operators.sqs`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.aws.operators.sqs import SQSPublishOperator  # noqa
+from airflow.providers.amazon.aws.operators.sqs import SQSPublishOperator  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.aws.operators.sqs`.",
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.sqs`.",
     DeprecationWarning,
     stacklevel=2,
 )
