+++ airflow/contrib/sensors/aws_sqs_sensor.py
@@ -17,15 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""This module is deprecated. Please use `airflow.providers.aws.sensors.sqs`."""
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.sqs`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.aws.sensors.sqs import SQSSensor  # noqa
+from airflow.providers.amazon.aws.sensors.sqs import SQSSensor  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.aws.sensors.sqs`.",
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.sqs`.",
     DeprecationWarning,
     stacklevel=2,
 )
