+++ airflow/providers/amazon/aws/operators/athena.py
@@ -21,7 +21,7 @@
 from uuid import uuid4
 
 from airflow.models import BaseOperator
-from airflow.providers.aws.hooks.athena import AWSAthenaHook
+from airflow.providers.amazon.aws.hooks.athena import AWSAthenaHook
 from airflow.utils.decorators import apply_defaults
