+++ airflow/providers/amazon/aws/sensors/athena.py
@@ -18,7 +18,7 @@
 # under the License.
 
 from airflow.exceptions import AirflowException
-from airflow.providers.aws.hooks.athena import AWSAthenaHook
+from airflow.providers.amazon.aws.hooks.athena import AWSAthenaHook
 from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.utils.decorators import apply_defaults
