+++ docs/autoapi_templates/index.rst
@@ -72,9 +72,9 @@ All operators are in the following packages:
 
   airflow/gcp/sensors/index
 
-  airflow/providers/aws/operators/index
+  airflow/providers/amazon/aws/operators/index
 
-  airflow/providers/aws/sensors/index
+  airflow/providers/amazon/aws/sensors/index
 
   airflow/providers/google/cloud/operators/index
 
@@ -106,7 +106,7 @@ All hooks are in the following packages:
 
   airflow/gcp/hooks/index
 
-  airflow/providers/aws/hooks/index
+  airflow/providers/amazon/aws/hooks/index
 
   airflow/providers/google/cloud/hooks/index
