+++ scripts/ci/pylint_todo.txt
@@ -190,8 +190,8 @@
 ./airflow/operators/slack_operator.py
 ./airflow/operators/sqlite_operator.py
 ./airflow/plugins_manager.py
-./airflow/providers/aws/operators/athena.py
-./airflow/providers/aws/sensors/athena.py
+./airflow/providers/amazon/aws/operators/athena.py
+./airflow/providers/amazon/aws/sensors/athena.py
 ./airflow/sensors/base_sensor_operator.py
 ./airflow/sensors/external_task_sensor.py
 ./airflow/sensors/hdfs_sensor.py
